requirements.txt
annotated-types==0.6.0
blis==0.7.11
catalogue==2.0.10
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
cloudpathlib==0.16.0
confection==0.1.4
contourpy==1.2.0
cycler==0.12.1
cymem==2.0.8
Cython==0.29.36
fonttools==4.50.0
idna==3.6
imageio==2.34.0
implicit==0.7.2
importlib_resources==6.4.0
Jinja2==3.1.3
joblib==1.3.2
kiwisolver==1.4.5
langcodes==3.3.0
lazy_loader==0.3
lime==0.2.0.1
MarkupSafe==2.1.5
matplotlib==3.8.3
mlxtend==0.23.1
mpmath==1.3.0
murmurhash==1.0.10
networkx==3.2.1
nltk==3.8.1
numpy==1.23.5
packaging==24.0
pandas==2.2.1
pillow==10.2.0
powerlaw==1.5
preshed==3.0.9
pydantic==2.6.4
pydantic_core==2.16.3
pyparsing==3.1.2
python-box==7.1.1
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.1
regex==2023.12.25
requests==2.31.0
scikit-image==0.22.0
scikit-learn==1.4.1.post1
scipy==1.12.0
six==1.16.0
smart-open==6.4.0
spacy==3.7.4
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.4.8
thinc==8.2.3
threadpoolctl==3.4.0
tifffile==2024.2.12
torch==2.3.0
tqdm==4.66.2
typer==0.9.4
typing_extensions==4.10.0
tzdata==2024.1
urllib3==2.2.1
wasabi==1.1.2
weasel==0.3.4
zipp==3.18.1
seaborn==0.13.2
tensorflow==2.16.1
umap-learn==0.5